fix(native): restore LIVE-tracking detection wash + text ON TOP — rip out baked-range/quiesce/hide/miss-grace gating (#1067)
Owner definitive P0: roll the wash back to the ORIGINAL live-tracking behavior
and change ONLY the paint order so glyphs render ON TOP (undimmed). This
supersedes the whole #1045/#1053/#1060/#1062/#1064 saga, which solved the wrong
problem by HIDING the wash.

End state — exactly two properties:
- LIVE-TRACKING: the behind-glyph wash carries the persistent #767 anchor's
  ABSOLUTE rows; HighlightPainter maps `viewRow = absRow - viewportOffset`
  against the SAME painted offset the glyphs use each frame, so it lands
  wherever the anchor CURRENTLY is — through scroll AND content churn, with no
  suppression gate and no baked/cached screen range. Removing the gating +
  grace-withholding makes the per-frame mapping the live resolution.
- TEXT ON TOP: paint order is background → highlight → text (already behind
  glyphs); pinned by paint_order_1045_test. Glyphs stay full-contrast.

Ripped out (the machinery that broke it):
- Quiesce-gate (#1064): washSuppressed, contentSettling / _washContentSettling,
  washHiddenForContentChurn.
- Hide-on-scroll (#1062/#1063): the isScrolling→suppress wiring, washHiddenForScroll.
- Miss-grace wash suppression (#1060): the in-grace wash withholding (both the
  resolver and no-resolver branches of _styledHighlights) + washSuppressedForGrace.
Kept: #1044 scan-perf cache + isScrolling scan-gating (scans still deferred; only
PAINT resolves live), the #985 single-damage-consumer contract (no new
RenderState), gutter chips, #1053 calm intensity.

Verification:
- New fork widget test wash_live_tracking_1067_test.dart: per-frame lockstep
  (`debugWashViewRows == { absRow - paintedOffset }`, never hidden) through a
  fling (60/60 frames) and streaming churn (62/80 frames). Full fork suite green.
- On-emulator wash_live_tracking_1067_shot_test.dart: same per-frame lockstep on
  the real device paint stack through scroll AND churn (exit 0), + mid-scroll and
  mid-churn screenshots showing the wash on its moving token with legible text on
  top. New debugWashViewRows test seam replaces debugWashSuppressed.
- Removed the now-dead gating tests (#1060/#1062/#1064 unit + shot).

Closes #1067
