+++ native/test/ui/detection_style_seam_test.dart
@@ -47,10 +47,6 @@ class _FakeGutterController extends ChangeNotifier
 
   @override
   bool get isScrolling => false;
-
-  @override
-  bool get contentSettling => false;
-
   @override
   Listenable get decorationListenable => this;
