+++ native/test/ui/ghostty_file_url_994_test.dart
@@ -68,10 +68,6 @@ class _FakeController extends ChangeNotifier implements TerminalController {
 
   @override
   bool get isScrolling => false;
-
-  @override
-  bool get contentSettling => false;
-
   @override
   Listenable get decorationListenable => this;
