+++ native/test/ui/ghostty_gutter_command_test.dart
@@ -77,10 +77,6 @@ class _FakeController extends ChangeNotifier implements TerminalController {
 
   @override
   bool get isScrolling => false;
-
-  @override
-  bool get contentSettling => false;
-
   @override
   Listenable get decorationListenable => this;
