+++ native/test/ui/ghostty_gutter_layer_test.dart
@@ -93,10 +93,6 @@ class _FakeController extends ChangeNotifier implements TerminalController {
 
   @override
   bool get isScrolling => _scrolling;
-
-  @override
-  bool get contentSettling => false;
-
   @override
   Listenable get decorationListenable => this;
